mxc_v4l2_capture: always call mxc_streamoff on last close
It makes no sense to check PIDs when the last close is
called, because the match may have occurred before the
last one was closed. In which case, it will never match.

Found when trying to get ov5642/ov5640_mipi cameras
working on test rig. Before this patch I could not
switch between cameras.

Signed-off-by: Troy Kisky <troy.kisky@boundarydevices.com>