gpu-viv: skip gpu dump as a workaround
To be re-visited, but on 8MP a GPU dump appears at boot-up when have
vivante kernel built-in with the Android branch.
This issue doesn't exist using the Linux branch (5.10) so something in
the Android tree changes the DMA sequence, for now skip the dump as 1)
we don't use them and 2) it doesn't prevent the GPU from recovering and
working ok.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>